refactor(solver): small naming changes and iterative search
