+++ pkg/controller/registry/resolver/solver/solve_test.go
@@ -333,3 +333,9 @@ func TestSolve(t *testing.T) {
 		})
 	}
 }
+
+// TODO: the search tests should verify that by the time we're done searching and returned
+// to the CardSort, we're no longer in a deeper test context than we should be.
+//func TestSearch(t *testing.T) {
+//
+//}
